myPage user정보 수정/조회 -> resume에서 user로 이동
