+++ albaing-api/src/main/java/com/jobjob/albaing/controller/ResumeController.java
@@ -7,20 +7,20 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping("/api")
+@RequestMapping("/api/resume")
 public class ResumeController {
 
     @Autowired
     private ResumeServiceImpl resumeService;
 
     // 1. 회원가입 시 이력서 자동 생성 (POST)
-    @PostMapping("/resume/create")
+    @PostMapping("/create")
     public void createResumeForUser(@RequestBody User user) {
         resumeService.createResumeForUser(user);
     }
 
     // 2. 이력서 수정
-    @PutMapping("/resume/update/{resumeId}")
+    @PutMapping("/update/{resumeId}")
     public void updateResume(
         @PathVariable int resumeId,
         @RequestBody ResumeUpdateRequest resumeUpdateRequest) {
@@ -31,23 +31,10 @@ public class ResumeController {
     }
 
     // 3. 이력서 조회
-    @GetMapping("/resume/{resumeId}")
+    @GetMapping("/{resumeId}")
     public Resume resumeDetails(@PathVariable int resumeId) {
         return resumeService.resumeDetails(resumeId);
     }
 
-    // 4. 사용자 정보 조회
-    @GetMapping("/user/{userId}")
-    public User getUserById(@PathVariable int userId) {
-        return resumeService.getUserById(userId);
-    }
 
-    // 5. 사용자 정보 수정
-    @PutMapping("/user/update/{userId}")
-    public void updateUser(@PathVariable int userId,
-                           @RequestParam("userEmail") String userEmail,
-                           @RequestParam("userAddress") String userAddress,
-                           @RequestParam("userProfileImage") String userProfileImage) {
-        resumeService.updateUser(userId, userEmail, userAddress, userProfileImage);
-    }
 }
