+++ albaing-api/src/main/java/com/jobjob/albaing/controller/UserController.java
@@ -1,20 +1,33 @@
 package com.jobjob.albaing.controller;
 
 import com.jobjob.albaing.dto.User;
+import com.jobjob.albaing.service.ResumeServiceImpl;
 import com.jobjob.albaing.service.UserServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping("/api/account/auth")
+@RequestMapping("/api/auth")
 public class UserController {
 
+    @Autowired
+    private UserServiceImpl userService;
 
 
 
+    // 마이페이지 - 사용자 정보 조회
+    @GetMapping("/user/{userId}")
+    public User getUserById(@PathVariable int userId) {
+        return userService.getUserById(userId);
+    }
 
+    // 마이페이지 - 사용자 정보 수정
+    @PutMapping("/user/update/{userId}")
+    public void updateUser(@PathVariable int userId,
+                           @RequestParam("userEmail") String userEmail,
+                           @RequestParam("userAddress") String userAddress,
+                           @RequestParam("userProfileImage") String userProfileImage) {
+        userService.updateUser(userId, userEmail, userAddress, userProfileImage);
+    }
 
 }
