+++ albaing-api/src/main/java/com/jobjob/albaing/service/ResumeService.java
@@ -9,12 +9,6 @@ public interface ResumeService {
         // 회원가입 시 이력서 생성
         void createResumeForUser(User user);
 
-        // 사용자 정보 조회
-        User getUserById(int userId);
-
-        // 사용자 정보 수정
-        void updateUser(int userId, String userEmail, String userAddress, String userProfileImage);
-
         // 이력서 조회
         Resume resumeDetails(int resumeId);
