+++ albaing-api/src/main/java/com/jobjob/albaing/service/UserServiceImpl.java
@@ -1,12 +1,28 @@
 package com.jobjob.albaing.service;
 
+import com.jobjob.albaing.dto.User;
+import com.jobjob.albaing.mapper.UserMapper;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 
 @Service
 public class UserServiceImpl implements UserService {
 
+    @Autowired
+    UserMapper userMapper;
 
+    // 사용자 정보 조회
+    @Override
+    public User getUserById(int userId) {
+        return userMapper.getUserById(userId);
+    }
+
+    // 사용자 정보 수정
+    @Override
+    public void updateUser(int userId, String userEmail, String userAddress, String userProfileImage) {
+        userMapper.updateUser(userId, userEmail, userAddress, userProfileImage);
+    }
